# Optimize Gelu by using reduced-range exp in erf approximation