Speed up `Erf` and `Gelu` by using the reduced-range exp(x) from https://github.com/robertknight/rten/pull/682 in the erf approximation. I re-ran `test_erf_exhaustive` to verify the max error is still within tolerance.

```
cargo test -p rten-vecmath -r -- --ignored --nocapture bench_erf

Before: reference 320510 us vectorized 71221 us. reference / vectorized ratio 4.500
After: reference 335342 us vectorized 61125 us. reference / vectorized ratio 5.486
```